Completed Knight Enemy Attack
Knight Attacks player by creating a collider in the space in front of
the knight sprite and then applying damage to player on trigger. Also
created a simple poison mushroom prefab, and cleaned up the way some
collisions with bullets and players and enemies worked. Also refined the
way health is managed, there is now a simple applyDamage function in
every player and enemy object. The applyDamage function is then accessed
by sending a message to the appropriate game object at the appropriate
time.